feat: Update Drizzle ORM and Kit setup content
Refactors the content for Drizzle ORM and Drizzle Kit setup. The previous content was too generic. The new content focuses on introducing the tools and assumes prior knowledge of `expo-sqlite`, then proceeds to guide the user through setting up a project from an Expo template. This provides a more structured and practical approach to learning.